src/third_party/blink/web_tests/external/wpt/mixed-content/gen/sharedworker-classic-data.meta/unset/xhr/same-http.keep-scheme.https-expected.txt

This is a testharness.js-based test.
FAIL Mixed-Content: Expects blocked for xhr to same-http origin and keep-scheme redirection from https context. assert_equals: The resource request should be 'blocked'. expected "blocked" but got "allowed"
Harness: the test ran to completion.
